requirements.txt
Django
djangorestframework
django-filter
django-extensions
django-easy-pdf
django-utils-six